# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Active Impact Talent Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Founder signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailVerificationInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [auth]
      summary: Resend verification email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List anonymized candidates (opted-in only)
      parameters:
        - in: query
          name: roleCategory
          schema:
            type: string
        - in: query
          name: seniority
          schema:
            type: string
        - in: query
          name: location
          schema:
            type: string
        - in: query
          name: openToRelocation
          schema:
            type: boolean
        - in: query
          name: yearsExpMin
          schema:
            type: integer
        - in: query
          name: yearsExpMax
          schema:
            type: integer
        - in: query
          name: compMin
          schema:
            type: integer
        - in: query
          name: compMax
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /candidates/breakdown:
    get:
      operationId: getCandidateBreakdown
      tags: [candidates]
      summary: Get candidate counts by role category and seniority
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateBreakdown"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get single candidate (anonymized for founders, full for admins)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    CurrentUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        company:
          type: ["string", "null"]
        role:
          type: string
          enum: [founder, admin]
        emailVerified:
          type: boolean
      required:
        - id
        - email
        - name
        - role
        - emailVerified
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/CurrentUser"
        message:
          type: string
      required:
        - user
    SignupInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        name:
          type: string
        company:
          type: string
      required:
        - email
        - password
        - name
        - company
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    EmailVerificationInput:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    ResendVerificationInput:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    Candidate:
      type: object
      properties:
        id:
          type: integer
        anonymizedHeadline:
          type: string
        roleCategory:
          type: string
          enum: [Engineering, Sales, Operations, Product, Finance, Marketing, Executive]
        seniority:
          type: string
          enum: [IC, Manager, Director, VP, C-level]
        yearsExperience:
          type: integer
        location:
          type: string
        openToRelocation:
          type: boolean
        compRangeMin:
          type: integer
        compRangeMax:
          type: integer
        topSkills:
          type: array
          items:
            type: string
        summaryBlurb:
          type: string
        notableCredentials:
          type: string
        status:
          type: string
          enum: [opted_in, paused, placed, withdrawn]
        dateAdded:
          type: string
          format: date
        hasRequestedIntro:
          type: boolean
      required:
        - id
        - anonymizedHeadline
        - roleCategory
        - seniority
        - yearsExperience
        - location
        - openToRelocation
        - compRangeMin
        - compRangeMax
        - topSkills
        - summaryBlurb
        - notableCredentials
        - status
        - dateAdded
        - hasRequestedIntro
...
    CandidateBreakdown:
      type: object
      properties:
        byRoleCategory:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
            required:
              - category
              - count
        bySeniority:
          type: array
          items:
            type: object
            properties:
              seniority:
                type: string
              count:
                type: integer
            required:
              - seniority
              - count
        total:
          type: integer
      required:
        - byRoleCategory
        - bySeniority
        - total
...